+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # 版本日志
 
+## 1.2.21
+
+*  【common-boot】 session 过期时间单位恢复为秒
+*  【common-boot】 参数验证器新增 或者判断条件
+
+---------------------------------------------------------------------
+
 ## 1.2.20
 
 *  【common-boot】 系统添加预置参数方法 cn.jiangzeyin.common.ApplicationBuilder.put 一般用于多模块间程序启动参数